Location restricted to abrantes only
With these minor changes, the map has now a boundary to city of Abrantes.

The animation speed needs a bit of tweaking, but still don't know if that is possible. Will check leaflet.js docs.

Also, the boundary was eye estimated, so that will also need a bit of tweaking.

Last but not least, map zooming was changed in order to the map stay focused on Abrantes and Abrantes only.