Dynamic errors (multi-lang) and Change from Globalization to Localization
My bad guys! It was supposed to be Localization from the beggining, as localization and globalization are two completely different things.

Anyways. the application in its entireness should be with multi language support, currently existing pt-PT and en-GB.

**Translations were brought to you by ChatGPT as im lazy to re-write all the text myself (plus me ingrish veri bad)**

Aight guys, probably no one will ever see this, but if you do... here's a cookie, mate! :cookie: