Settings page w/ language selector
Created a settings page with a selector so it is possible to select a language from the available languages (JSON file).

The icon that is in the settings page is the same as the augmented reality page (for now).

The language selector is very simple. I am open to discuss future UI for the select.